Fix nested templates and PAPI chaining
