--- .github/workflows/run-fix302-round6.yml
@@ -1,25 +1,0 @@
-name: Run Codex round six fixes
-on:
-  push:
-    branches:
-      - security/javascript-placeholder-bindings
-permissions:
-  contents: write
-jobs:
-  patch:
-    if: github.actor != 'github-actions[bot]'
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: security/javascript-placeholder-bindings
-      - name: Apply fixes
-        run: python3 .github/fix302-round6.py
-      - name: Commit fixes
-        run: |
-          rm -f .github/fix302-round6.py .github/workflows/run-fix302-round6.yml
-          git config user.name 'Ben'
-          git config user.email 'benbergen12@gmail.com'
-          git add -A
-          git commit -m 'Fix nested templates and PAPI chaining'
-          git push origin HEAD:security/javascript-placeholder-bindings
